added default.txt sample to sdcard

+++ src/customize.sh
@@ -5,21 +5,29 @@
 sdcard_folder="/sdcard/DynamicDetachX"
 # setup new sdcard dir
 [ ! -d "$sdcard_folder" ] && {
-    ui_print "- setting up sdcard folder"
+    ui_print "* setting up sdcard folder"
     mkdir -p "$sdcard_folder"
     cp -rf "$MODPATH/detach.txt" "$sdcard_folder/detach.txt"
 }
 
+
 # preserve the old Detach.txt
 old_detach_txt="/data/adb/modules/dyndetachx/detach.txt"
-[ -f "$old_detach_txt" ] && {
-    ui_print "- updating old detach.txt"
+if [ -f "$old_detach_txt" ]; then 
+    ui_print "* using old detach.txt"
     cp -rf "$old_detach_txt" "sdcard/DynamicDetachX/detach_old.txt"
     cp -rf "$old_detach_txt" "$MODPATH/detach.txt"
     set_perm "$MODPATH/detach.txt" root root 0755
-}
+elif [ -f "$sdcard_folder/detach.txt" ]; then 
+    ui_print "* using sdcard detach.txt"
+    cp -rf "$sdcard_folder/detach.txt" "$MODDIR/detach.txt"
+    set_perm "$MODPATH/detach.txt" root root 0755
+else
+    ui_print "* creating $sdcard_folder/detach.txt"
+    printf "com.google.android.apps.youtube.music\ncom.google.android.youtube" > "$sdcard_folder/detach.txt"
+fi
 
 # check for dynmount
 [ ! -d "/data/adb/modules/magisk_proc_monitor" ] && {
-    ui_print "Missing Magisk Process Monitor. Please install it after installing this Module."
+    ui_print "* missing Magisk Process Monitor. Please install it after this module."
 }
